+++ Riot/Managers/Settings/RiotSettings.swift
@@ -56,6 +56,7 @@ final class RiotSettings: NSObject {
         static let homeScreenShowPeopleTab = "homeScreenShowPeopleTab"
         static let homeScreenShowRoomsTab = "homeScreenShowRoomsTab"
         static let homeScreenShowCommunitiesTab = "homeScreenShowCommunitiesTab"
+        static let unifiedSearchScreenShowPublicDirectory = "unifiedSearchScreenShowPublicDirectory"
     }
     
     static let shared = RiotSettings()
@@ -447,4 +448,18 @@ final class RiotSettings: NSObject {
             defaults.set(newValue, forKey: UserDefaultsKeys.settingsSecurityScreenShowAdvancedUnverifiedDevices)
         }
     }
+    
+    // Mark: - Unified Search
+    
+    var unifiedSearchScreenShowPublicDirectory: Bool {
+        get {
+            guard defaults.object(forKey: UserDefaultsKeys.unifiedSearchScreenShowPublicDirectory) != nil else {
+                return BuildSettings.unifiedSearchScreenShowPublicDirectory
+            }
+            return defaults.bool(forKey: UserDefaultsKeys.unifiedSearchScreenShowPublicDirectory)
+        } set {
+            defaults.set(newValue, forKey: UserDefaultsKeys.unifiedSearchScreenShowPublicDirectory)
+        }
+    }
+    
 }